==> 55_Entregable tabulados/rutinas/03_2_cites_tablas.R <==
rm(list = ls())

library(tidyverse)
library(rio)

bdd <- readRDS("55_Entregable tabulados/intermedios/03_cites.rds")

t2 <- bdd |> 
  group_by(tipo, unidades, anio) |> 
  summarise(cantidad = sum(as.numeric(cantidad), na.rm = T),
            casos = n())
